[WIP] add modelfitter for emission line data
The idea is to use pymc to fit a model to a spectroscopic time-series
dataset of time X velocity X flux.  The data are approximated as
multiple emission lines varying (in velocity) sinusoidally, with the
flux for each line core being drawn from a gaussian.

The fitting is working on some mock test data; yet to deploy on real
data.